Change: Move almost all of onQRScan to event
